[ATC] Update user doc.
The target function call to which a `let` or `mv-let` is bound does not
necessarily have to be a recursive one (representing a loop): it can also be a
non-recursive one, representing a void function call used as an expression
statement. This has been allowed for a while by ATC, but apparently the user
documentation had not been updated in this respect.

Thanks to Grant Jurgensen and Eric Smith for pointing out that the documentation
did not allow a non-recursive function call in these circumstances.