[C deep] Improve dynamic semantics.
Make `exec-arrsub-of-member` consistent with the combination of `exec-member`
and `exec-arrsub`, so that we can then streamline the definition of the dynamic
semantics and use `exec-arrsub-of-member` as a derived operation used in the
symbolic execution.

Some additions and adaptation to the symbolic execution were needed, including
in the struct-specific theorems generated.