[X86ISA] Avoid `subset-equal`.
Before this commit, a file in the X86ISA project defined a `subset-equal`
function that is similar to the built-in `subsetp-equal` except for a slightly
weaker guard. This commit removes that function, using the built-in
`subsetp-equal` instead. A guard of another function had to be slightly
strengthened so that its call of `subsetp-equal` can be guard-verified.