[ATC] Rename two `defstruct`-generated operations.
struct-<tag>-read-<member>-all  ==> struct-<tag>-read-<member>-list
struct-<tag>-write-<member>-all ==> struct-<tag>-write-<member>-list

These are the operations that return the list of all the elements of an array
member of a struct. The unqualified name `struct-<tag>-read/write-<member>` is
not used for these, because that should more reasonably return the actual array
member, which is an array and not a list.