[ATC] Simplify and speed up some proofs.
These are the proofs of the symbolic execution rules that are struct-specific
and are generated when ATC is called. The rules for the readers of elements of
array members that take specific C integer types for the indices are now proved
easily and more quickly from the newly generated theorems for the more general
readers that take all C integer types.