[ATC] Use tau in a generated theorem.
Adding type prescriptions for the new struct operations, as done for similar
operations previously, didn't seem to work in at least one example. This
generated theorem will be eventually replaced by the more granular theorems that
are being now partially generated, but in the meanwhile we add tau to the
enabled theory, which works on all our current examples. This should be robust
in the sense that the tau database should never shrink and it should never turn
something provable into something unprovable.