[ATC] Add a check on the code.
We now require the two branches (i.e. 2nd and 3rd argument) of a `c::condexpr`
to not have C types of rank lower than int (i.e. no unsigned or signed char or
short). The reason is that otherwise the conditional expression performs an
implicit conversion on the branches, but in the ACL2 representation of C code
handled by ATC all conversions must be explcit.

Thanks to Eric Smith for running ATC on an example that resulted in a failing
well-formedness (i.e. static semantic checks) theorem, due to (1) C making the
conversions but (2) ATC not having the convertsions.