Made more iprinting mods, including a change to read-object.
Quoting :DOC note-8-6, expanding an existing paragraph:

  Arranged that [iprinting] that takes place during [break-rewrite] or
  [cw] is better reflected outside break-rewrite.  For examples, see
  [community-books] input file
  books/system/tests/iprint-and-brr-input.lsp, which contains
  comments on problematic interactions between break-rewrite and
  iprinting in Version 8.5, and which generates (via the [run-script]
  utility) the output in file iprint-and-brr-log.txt in that
  directory.  Note that [read-object] now updates data structures so
  that it can read back in what was just printed by [cw]; see a
  comment in the definition of read-object in the ACL2 source code.

Logically, read-object updates the ACL2 oracle to reflect iprinting
changes.  So read-object no longer leaves the state unmodified.  Among
the several changes to the books is that the lemma
state-preserved-by-read-object-when-eof in books/std/io/base.lisp has
been commented out, because it is no longer true.