[ATC] Extend some code for modular proofs.
Add code to construct the scope (alist) for global variables (in static storage)
for the initial symbol table of a C function. These are the global variables
passed to the ACL2 function that are declared via `defobject`.

Differentiate the code to construct the scope (alist) for local variables (in
automatic storage) for the initial symbol table of a C function. These are the
parameters of the function (excluding the `defobject` ones).