[ATC] Extend modular proof generation.
Handle `mbt` and `mbt$`.

Even though currently we can't stop ACL2 from expanding `return-last`, it seems
that the modular proofs work if that is consistently expanded.

This required a little additional work because of the possibility of embedded
`if`s in `mbt` tests (either directly, or indirectly via `mbt$`). All the
current uses of ATC work, but we also need to perform the `if`-`if*`
substitution in the `else` branches.

We will also streamline things a bit, as it is only necessary to support `mbt`
directly, given that `mbt$` expands to it and that ATC operates on translated
terms.