[ATC] Streamline and improve support for `mbt`.
Since `mbt$` expands to `mbt`, and ATC operates on translated terms, there is no
need to expliictly support `mbt$`; `mbt` suffices. The user documentation and
the implementation have been updated accordingly.

Make the modular proofs generated for `mbt` more robust by replacing `if` with
`if*` not only in the argument of the `mbt` test, but also in the 'else' branch.

Simplify some code.