[ATC] Extend modular proofs.
Add handling of the empty lists of block items that may follow a variable
declaration or assignment, so that the `cons` of a list of block items can
be handled uniformly.

Add a few more temporary checks to stop modular proof generation under certain
conditions, now that the extension mentioned in the previous paragraph take
modular proofs further than before.