[ATC] Extend code for modular proof generation.
This is code for the theorem about the execution of a call of a `void` function.
The event is not submitted yet because we need some extensions to some of the
existing modular theorems, having to do with the return types of the objects
side-effected by expressions and statements and block items and functions.