[ATC] Adjust handling of called function limits.
Avoid substituting actuals for formals, because the term checker that returns
the limit term already performs all the necessary substitutions, taking
(essentially) the context into account. This made no difference so far, because
limits are always constant without loops (which are not yet supported in modular
proofs), but it is of course important when we extend modular proof generation
to loops.