[ATC] Extend a term checker.
This is for non-recursive function calls. We fold into that checker also the
checks on the arguments, and we add a missing check in the process, for
arguments of formals that represent external objects.

Next we will eliminate the separate function that just checks the arguments,
since it is now redundant, and also misses the check that we just added.