Disable some built-in printing functions (e.g., fmt).  Make an include-book local in books/system/fmt-support.lisp.  Modify books/kestrel/acl2data/gather/tests/ to avoid some unfortunate regression behavior.
Thanks to Eric Smith for suggesting the list of built-in printing
functions to disable and for suggesting making that include-book
local.  Thanks to Eric McCarthy for pointing out one of at least two
regression problems involving books/kestrel/acl2data/gather/tests/,
and to Eric Smith for contributing to the discussion; as far as I know
things are good there now.

Passed "make devel-check" and "make proofs" in addition to the usual
regression-everything.  [The devel-check was done before adding an
initial local in-theory to books/system/fmt-support.lisp, which
shouldn't matter because it's conditionalized on #-acl2-devel.]