[APT] Improve the `restrict` transformation.
Refine the proofs generated for the old-to-new theorem to use proof builder
instructions that should be more robust. It turns out that the change made in
the previous commit is still insufficient to make the inspiring example work,
due to the fact that, in that example, the restricting predicate equates a
formal with a quoted constant, and ACL2 replaces the former with the latter,
making the `:expand` no longer apply.

In the apparent absence of a simple way to tell ACL2 to always expand the
definitions once, no matter what, we achieve this (in a less simple way) via
suitable proof builder instructions that "inject" the expanded definitions into
the proofs so that it can be taken advantage of during the proof.

Another complication is that there may be multiple induction steps and base
cases, and we don't easily know in advance how many, so we use a `(:repeat
(:do-all ...))` to apply the instructions to all the induction subgoals.