Incorporated fix from J to a rare wormhole bug.
Thanks to Eric Smith for reporting the bug, with an example that I
used to construct the following simple example.

(table evisc-table 100 "one-hundred")
:brr t

J observed that the following is another example.

(wormhole-eval 'DEMO
               '(lambda (whs)
                  (prog2$ (cw "Bug: ~X01"
                              '(here is data)
                              (evisc-tuple 3 5 nil nil))
                          whs))
               nil)

In both cases ACL2 produced raw Lisp errors; now it doesn't.

J's fix is to function brr-evisc-tuple-oracle-update (and the @par
version) and is explained in a comment added to its defun.

There is no release note since this bug was introduced after
Version 8.5.