[X86ISA] Tweak models of `div` and `idiv`.
The width should be the same as the low bit index, i.e. 8 times the operand size
in bytes (i.e. the operand size in bits). It is actually equivalent, because the
extra bits are 0, but it seems cleaner to do it this way.

This issue was brought up by Eric Smith.

A guard verification lemma had to be tweaked in order for the guard proofs to go
through.