Tweaked :DOC df and :DOC soundness.  Updated the Essay on Support for Floating-point (double-float, df) Operations in ACL2, in particular by adding a new section, On overflow and soundness.  Added build-time checks about floating-point overflow.  Documented new supported value for (@ script-mode), skip-ldd-n.  Fixed :DOC case.  Tweaked warning about modifying a user stobj.  Fixed typo in :DOC type and tweaked :DOC hints-and-the-waterfall.  Resynched :DOC topics acl2-doc and acl2-doc-summary.  Added :DOC topics for trace-co and (as pointer) set-trace-co, added these symbols to *acl2-exports*, and linked to trace-co in :DOC trace$.  Fixed verify-guards for lambda and lambda$ objects.
Thanks to Pete Manolios, Eric Smith, and J Moore for discussions
leading to the changes pertaining to floating point / df.  Mostly they
deal with the question of whether overflow might produce a non-df
answer such as "nan" -- in a nutshell, no.  The soundness claim in
:DOC df and :DOC soundness has been tightened up a bit to require not
just the same Lisp but also the same environment, as suggested by
Pete.

Thanks to Eric Smith for suggestions leading to the fixes to :DOC case.

For the warning about modifying a user stobj, the change is to report
"may have modified the user stobj" instead of "has modified the user
stobj".  To see the change you can evaluate the following two forms
and look at the output from the second.

(defstobj st fld)
(value-triple st :stobjs-out :auto)

Thanks to Warren Hunt for pointing out a typo in :DOC type, for
suggesting an earlier mention in :DOC hints-and-the-waterfall of the
settled-down process, and for a suggestion resulting in resynching of
:DOC topics acl2-doc and acl2-doc-summary.

Verify-guards was causing a raw Lisp error when applied to a
non-symbol.  That bug was introduced after v8-5 and is now fixed.