[X86ISA] Temporarily(?) fix an issue.
The model was throwing a #UD when a SHLX instruction was designating RAX via the
vvvv VEX bits, which should not be the case (a compiler generated such
instruction). RAX is encoded in one's complement form in the vvvv bits, so those
bits are 1111 in that case. The generated VEX dispatch code had a check to
ensure that the vvvv bits of SHLX (and SARX and SHRX) are never 1111, but that
check should be made only if the instruction does not use the vvvv bits, but
SHLX/SHRX/SARX do use them.

For now I've commented out a line that's responsible for generating that check
that vvvv is not 1111, but it may not be a general fix. So I put a 'TODO' there.