Guarantee that floating-point overflow always signals an error.
When the proof is attempted for lemma1 in the Example below, a
floating-point overflow occurs.  If that overflow is not detected,
then we can prove nil (as shown in the Example below), which was in
fact possible for ACL2 built on LispWorks or Allegro CL before this
commit.  Also, tightened up such code for GCL; thanks to Camm Maguire
for that suggestion.  Thanks again to Pete Manolios and Eric Smith for
questions relevant to all this during my Feb. 8 talk.

Implementation notes.  The approach to causing error on overflow is
two-fold.  Function break-on-overflow-and-nan, defined in source file
acl2.lisp, arranges that this occurs in GCL, CCL, SBCL, and CMUCL.
Macro df-signal?, defined in source file float-raw.lisp, is the
identity in those Lisps, but checks computation results for LispWorks
and Allegro CL (without too much of a hit on efficiency, I hope -- but
I don't know).

Example:

(defun bar (x1 x2)
  (declare (type double-float x1 x2))
  (let ((y (df* x1 x2)))
    (and (dfp y) (not (df= (df- y y) (df0))))))

(defthm lemma1
  (let ((x (to-df #d1.7976931348623157d+308)))
    (bar x x))
  :rule-classes nil)

(defthm lemma2
  (not (bar x x))
  :rule-classes nil)

(defthm bad
  nil
  :hints (("Goal"
           :use
           (lemma1
            (:instance lemma2
                       (x (to-df #d1.7976931348623157d+308))))))
  :rule-classes nil)

End of Example.