[X86ISA] Add writer for ZMM register or memory.
This is analogous to two functions, one to write to a general-purpose register
or to memory, and one to write to an XMM register (for a non-VEX instruction) or
memory. The new functions writes to a ZMM register or to memory; writing to a
ZMM register includes writing to a YMM register, and writing to a XMM register
for a VEX instruction.