Floating-point overflow is always detected.  Fixed bug in function declaims used by GCL.  Worked around a GCL bug in =.  Fixed a bug in how apply$ operates on lambdas with df formals.  Documented that a lambda$ body must return a single ordinary value, not a df.  Modified acl2-doc commands TAB and Shift-TAB to alert on wrapping.  Tweaked books/projects/vwsim/read-float.lisp to avoid GCL compiler error.  Replaced `::' by `:' in books/projects/numbers/triangular.lisp to avoid GCL error.
ACL2 now detects floating-point overflow in every Lisp implementation.
In Allegro CL and LispWorks, there was no way to tell Lisp to do this,
so we check the output against infinity (positive or negative) using
the macro df-signal? (which is the identity in the other Lisps that
can host ACL2).

Fixed a bug in source function output-type-for-declare-form-rec in the
case of ec-call and dfs, which is used by GCL to determine suitable
declaim forms for functions.

Worked around a GCL bug that seems to return t for (= x (to-df x)) for
double-float values of x.  (See source function make-df.)  This may
not be a complete workaround, but apparently this bug will disappear
in soon-to-be-released GCL 2.7.0.

Fixed a bug in how apply$ operates on @(tsee lambda) objects and
@(tsee lambda) expressions when there are @(see df) formals.  This
bug didn't show up in CCL or SBCL but it did in GCL, and I was able
to replicate it in CCL with safety 3.

Quoting :DOC note-8-6:

  In [ACL2-doc], modified the TAB and Shift-TAB (sometimes known as
  <backtab>) commands so that they alert the user when wrapping.