[acl2s] Replace ACL2s' PRNG with the Thrust PRNG (#1591)
I replaced the PRNG used by ACL2s with the Thrust PRNG, which is
designed to work with 63-bit state (the seed) and 64-bit outputs. My
implementation is an adaptation of Tommy Ettinger's C implementation,
which Tommy dedicated to the public domain. The implementation is here:
https://gist.github.com/tommyettinger/e6d3e8816da79b45bfe582384c2fe14a

Some quick-and-dirty benchmarking indicates that this implementation is
about 2x faster than the previous one.

I also needed to update some of the theorems that were being proved
about genrandom-seed and genrandom-state, as they often used `car` to
access the first element of the MV that those functions return. I
replaced `car` with `mv-nth 0` in those theorems, which eliminated
failures that were occurring in random-state.lisp.