[axe] Improve generated rewriters.
When we are to memoize, do 2 rewrites, first one with memoization and then a second one using node contexts (but no memoization, for soundness).  We sometimes need to use the node contexts but we also want memoization on when most of the work is being done (can make a huge difference in the speed of large rewrites).