[C$] Add initial version of parser.
This has not been tested or reviewed at all yet.

It covers all of C after preprocessing, but it currently has some documented
limitations in the handling of some complex inherently ambiguous syntactic
constructs that are notoriously complex in C. This will be remedied soon, but
the current parser should be a good starting point.