[c$] Create directory for tempfile in preprocess-file
The oslib::tempfile utility may return a filepath whose directory does not yet exist. preprocess-file now creates said directory before invoking the preprocessor.

This also removes the absolute-filepath utility, and simply uses unaltered relative filepaths instead. The original reason for making filepaths absolute was that it was unclear whether relative filepaths would always be interpreted with respect to the CBD. In practice, it seems this is in fact the case. Further, the absolute-filepath function failed to properly interpret special directory names such as ~.