Make tshell-call sound by defining in terms of read-acl2-oracle
The existing version of tshell-call is unsound because it does not take
nor return state, and is non-functional. See the Zulip discussion in
April:
https://acl2.zulip.kestrel.institute/#narrow/stream/19-general/topic/tshell.20unsoundness.

This fixes the issue by redefining tshell-call-fn1 in terms of
read-acl2-oracle (it is no longer a partial encapsulate). All uses of
tshell-call in the community books were updated to accomodate state. We
also introduce a new book, tshell-unsound, which maintains the unsound
stateless variants for convenience. Any users of tshell outside the
community books may use the definitions here if it is too burdensome to
refactor to accomodate the new state argument.