[C$] Improve abstract syntax.
Use the newly introduced fixtype for ambiguous expressions and type names in the
ambiguous `sizeof` case of expressions.

Adapt the parser accordingly. This is just so the parser certifies with the
change to the abstract syntax. Actually extending the parser to take advantage
of the enhanced abstract syntax is still to be done.

Adapt the printer accordingly. Since `sizeof(I)` with `I` identifier may be
common, we temporarily accept it in printing. Normally, semantic analysis should
eliminate ambiguous constructs, prior to printing. Establishing whether an
ambiguous expression or type name is an identifier takes a little bit of
computation, for which we introduce an operation.