[C$] Speed up parser.
Cache the parser state size in the parser state. This is the size of the
remaining input, including unread characters and tokens. Some profiling revealed
that significant time is spent in the function that calculates that size. That
function is needed to save the size prior to updating the parser state (in a
single-threaded way, to accommodate future stobjification) in order to support
`mbt` tests needed for parser termination in mutually recursive functions.

The caching of the size takes advantage of the `:require` feature of
`fty::defprod` to state the redundancy of the new state component and its
consistent update. Only few functions of the parser, those that manipulate the
parser state directly, had to be modified. The function that returns the size
has been changed to use an `mbe`, so that its logical definition does not
change, but its run-time execution is faster (it just retrieves the cached
size).

Thanks to Eric Smith for running the CCL profiler and reporting that significant
time was spent, on an example, in the aforementioned function that returns the
size.