[C$] Speed up parser.
Cache the number of read tokens in the parser state. Previously the code, as
part of the checkpointing and backtracking mechanism, had to take the `len` of
the `tokens-read` component of the parser state, which is the list of all the
tokens read so far. As these grow, the time to take the `len` grows. Now we
cache that length in the parser state, using `fty::defprod`'s `:require` feature
to enforce consistent updates, to speed up the retrieval of that length.

Thanks to Eric Smith for running the CCL profiler on the parser, which showed
that significant time was spent in the functions that took the `len` of that
parser state component.