[C$] Add support for adjacent string literals.
We extend the ABNF grammar to allow multiple adjacent string literals where only
one was previoulsy allowed: this captures adjacent string literals at the phrase
grammar level, instead of outside the grammar.

Correspondingly, we extend the ASTs to hold lists of string literals at the same
places, so that we preserve that information.

Extend parser and printer to handle adjacent string literals.