[C$] Add support for GCC extension `__builtin_va_list`.
This is a built-in type, not readily documented in the GCC manual, but observed
in practical code and verified in a GCC implementation.

Previously we were treating this as a typedef name in the disambiguator, which
was adequate for past purposes, but the correct treatment is that of a
(non-standard) type specifier.