Put cbd-fn in guard-verified logic mode.
Thanks to Eric Smith for requesting that cbd-fn be changed to call (er
hard? ...) instead of (er hard ...), so that it could be in logic mode
with guards verified.  So I made that change to the er call and then
put cbd-fn in guard-verified logic mode.

Passed "make proofs" in addition to "make regression-everything".