Two very technical changes: Tweaked implementation of attachable stobjs, and improved a build error message.
The following explanations are best ignored except by implementors.

The first change is to avoid writing notinline directives to expansion
files.  It takes advantage of the fact that functions are in the world
before they are compiled, and that they are proclaimed notinline when
necessary as they are put into the world (by add-trip's calls to
install-defs-for-add-trip).  This fact is relevant since it is
long-standing practice for declaims in expansion files to be guarded
by a readtime conditional restricting to the current host Lisp, and
that might have had to change in support of include-book option
:load-compiled-file :comp now that notinline directives are necessary
for correct evaluation.

The build error message that was improved was when (boundp
'COMMON-LISP::*PRINT-PPRINT-DISPATCH*) fails, to use
exit-with-build-error instead of exit-lisp.