Fixed a bug in applying useless-runes.
I found this bug using a safety-3 build of a very recent ACL2.  I'm
pretty sure that this bug goes back at least to Version 8.5, but I'm
not positive, so I made the following a comment in (deflabel note-8-6
...) rather than a release note item.

; A bug was fixed that was probably present in Version 8.5 (and perhaps earlier
; versions), and definitely present in a github version of ACL2 before the
; release of Version 8.6, that permitted an out-of-bounds array access when
; updating a theory to accommodate useless-runes.  That error could occur
; during a call of compress1 from update-enabled-structure, from
; load-theory-into-enabled-structure-1, from useless-runes-ens.  The latter two
; functions were fixed to avoid the error.  This error was found during a run
; with ACL2 built using safety 3.

Passed "make basic" at safety 3, in addition to the usual
regression-everything at normal (0) safety.