[c2c] Add specialize and copy-fn transformations
The specialize transformation is akin to partial application; it drops
a parameter from a function, and converts it into a declaration which is
initialized to a constant at the top of the function body.

The copy-fn transformation simply duplicates a function, changing the
name. It is only useful in combination with other transformations (e.g.
copying a function and then specializing the copy, leaving the original
intact). Right now, copy-fn does not appropriate rename recursive calls
in the copied function. This should change soon.