[C$] Fix parser bug.
This is related to supporting GCC statement expressions. It turns out that we
needed to take them into account not only when parsing primary expressions
(statement expressions are primary expressions), but also when parsing postfix
expressions; in addition, when parsing cast expressions, where we were already
taking statement expressions into account, after parsing those we need to parse
the rest of the postfix expression if any.