Stop x86isa virtualization debugging tool crash when RIP is unmapped
Previously, if the RIP can't be translated (which may happen often on
operating systems which do demand paging) the debugging tool would
segfault since it tried to access invalid memory because of an erroneous
computation for where the data at the instruction pointer is.

Now, we avoid looking for and printing the offending fields if the RIP
translation fails.