Eliminated macro for include-book in recently-added directory, books/projects/fm9801/.  Fixed numerous typos in comments.  Avoid printing "Hons-Note" messages by run-script.
The macro for include-book, include-acl2-book, was defeating
dependency analysis.

Thanks to Bob Boyer for reporting the typos, which he found using
gmail's compose feature.

During a GCL-based regression, a "Hons-Note" message was printed that
I hadn't seen before, causing a failure in a run-script book.  Now
such messages are suppressed, in analogy to time-tracker messages.