Use query params for xkey in SEO friendly xdoc
The SEO friendly version of xdoc put the xkey in the path component of
the url (i.e. URLs would look like `index-seo.php/ACL2___TOP`). While
this is more idiomatic than having them be a query parameter, this
caused mod_evasive to recognize each page as separate and thus ratelimit
them separately instead of together.

This moves the xkey to an `xkey` query parameter and updates the URLs
accordingly.