Made another (final?) fix to help dependency analysis.
As Eric Smith pointed out, when my previous fix to
books/kestrel/acl2data/gather/tests/cert.acl2 eliminated a linebreak
in a LD call, it restored a problem with dependency analysis since the
pathname referenced a directory that might not yet have been created.
Comments in that file explain the problem and the new solution.