[c$] Return msgp errors in preprocess-file
The preprocess-file function now returns a message instead of triggering
a soft error. The signature of the function was also changed (the
adoption of the error-value-tuples utilities made it more convenient to
add more returns). The signature of preprocess-files is unchanged.

This also changes the tests file accordingly.

Finally, the stdout/stderr output of the preprocessor is now recorded
and returned with the error message on failure.