[C$] Extend disambiguator for a GCC built-in.
We add to the initial disambiguation table a built-in function that is not
actually documented in the GCC manual, which says that built-ins related to
variable-length arguments (as this one is) are not explicitly documented because
they are subject to change and should not be used. We are seeing those probably
just because the C preprocessing is exposing them, by expanding definitions of
GCC-implementation-dependent macros that are legitimate to use.