[C$] Add support for the GCC `__stdcall` extension.
This is a bit odd, because `stdcall` is an attribute, which should therefore be
written as `__attribute((stdcall))` or `__attribute__((stdcall))`. However, we
have encountered it in practical code written as just `__stdcall`.  Experiments
suggest that this is not just a shortcut for the longer syntax, because it is
not allowed in places where the normal attribute syntax is allowed. It appears
to be its own special thing. We could not find this special syntax documented in
the GCC manual.

This commit adds support for this construct, as a possible declaration
specifier.

Add a test to demonstrate support.

Thanks to Eric Smith for running the tools on some code that uses this
construct.