[C$] Fix bug in disambiguator.
When ambiguous cast/binary expressions are disambiguated to binary expressions,
they may decrease their priority from the one of cast expressions; the latter
arises from the parser. Thus, depending on which super-expressions they occur
in, it may be necessary to re-arrange the super-expressions to avoid changing
the meaning of the expressions.

Thanks to Eric Smith for running the disambiguator on an example that exposed
the issue and led to this fix.